backend/requirements.txt
flask==3.1.0
flask-restx==1.3.0
flask-jwt-extended==4.7.1
flask-caching==2.3.0
flask-limiter==3.8.0
flask-cors==5.0.0
redis==5.2.1
python-slugify==8.0.4
bleach==6.2.0
markdown==3.7.0
bcrypt==4.2.1
marshmallow==3.23.2
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
